requirements.txt
anyio==4.11.0
APScheduler==3.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
fragment-api-lib==1.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psutil==7.1.0
PyMySQL==1.1.2
python-dotenv==1.1.1
python-telegram-bot==22.5
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
